requirements.txt
streamlit
requests
beautifulsoup4
google-api-python-client
google-auth
